+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - New functions `uptake_pct_govtarget_school2()` and `uptake_pct_godutch_school2()`
 - Enables download of other OD datasets with `get_od()` (#66)
+- New argument `verbose`, false by default, in uptake functions
 
 # pct 0.5.0